% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_obs_noise}
\alias{add_obs_noise}
\title{Add observational noise}
\usage{
add_obs_noise(
  df,
  X_names,
  noise_mean = 0,
  noise_sigma = 0.01,
  noise_constant = 0,
  seed_nr = NULL
)
}
\arguments{
\item{df}{Dataframe}

\item{X_names}{Names of columns in df to add observational noise to}

\item{noise_mean}{Mean of white noise}

\item{noise_sigma}{Standard deviation of white noise}

\item{noise_constant}{Constant to add to observed timeseries}

\item{seed_nr}{Seed number for generating noise}
}
\value{
Noisy dataframe
}
\description{
Add observational noise
}
