% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_par_as_cols}
\alias{add_par_as_cols}
\title{Add parameters in list as columns in dataframe}
\usage{
add_par_as_cols(df, for_par)
}
\arguments{
\item{df}{Dataframe}

\item{for_par}{List of parameters}
}
\value{
Updated dataframe
}
\description{
Add parameters in list as columns in dataframe
}
