% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{apply_filter_regime_switches}
\alias{apply_filter_regime_switches}
\title{Apply filter to obtain desired regime switches}
\usage{
apply_filter_regime_switches(
  regime_bounds_df,
  regime_switch_list,
  trans_or_null = c(NA, "null", "transition")[1]
)
}
\arguments{
\item{regime_bounds_df}{Dataframe with regime bounds concatenated of multiple simulations}

\item{regime_switch_list}{List of filters per regime switch, output of make_filter_regime_switches()}

\item{trans_or_null}{Transition or null model}
}
\value{
Filtered regime bounds dataframe
}
\description{
Apply filter to obtain desired regime switches
}
