% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{bifurcation_ts}
\alias{bifurcation_ts}
\title{Generate timeseries with changing bifurcation parameter}
\usage{
bifurcation_ts(
  model,
  model_pars,
  bifpar_list = NULL,
  bifpar_pars = NULL,
  X0 = c(),
  X_names = names(X0),
  seed_nr = 123,
  timestep = 0.01,
  nr_timesteps = 100,
  deSolve_method = c("lsoda", "euler", "rk4")[1],
  stopifregime = function(out) {
     any(apply(out, 2, is.infinite)) | any(apply(out, 2,
    is.nan))
 },
  do_downsample = TRUE,
  fs = 50,
  which_X = "all",
  silent = FALSE,
  max_iter = 1,
  seed_change = function(seed_nr) {
     seed_nr + 1
 }
)
}
\arguments{
\item{model}{ODE in the style of deSolve}

\item{model_pars}{Fixed model parameters}

\item{bifpar_list}{List with changing bifurcation parameter values}

\item{bifpar_pars}{List with parameters to generate bifpar_list, needs to include list(bifpar_start = .96, bifpar_end = .96), optionally specify list(pre_steps = 0, baseline_steps = 100, transition_steps = 0, post_steps = 0)}

\item{X0}{Initial state}

\item{X_names}{Names of variables in model}

\item{seed_nr}{Seed number}

\item{timestep}{Step size in time}

\item{nr_timesteps}{Number of time steps}

\item{deSolve_method}{Method of generating ODE passed to deSolve::ode}

\item{stopifregime}{End generating timeseries if this function is satisfied}

\item{do_downsample}{Reduce dataframe size by downsampling?}

\item{fs}{Sampling frequency for downsampling if do_downsample = TRUE}

\item{which_X}{Time point(s) in downsampling window to select if do_downsample = TRUE}

\item{silent}{Don't output progress if TRUE}

\item{max_iter}{Maximum number of tries to restart from a new initial condition and not hit undesirable regime}

\item{seed_change}{Function to update seed number when restarting after hitting undesirable regime}
}
\value{
List with timeseries matrix and parameters used to generate the timeseries
}
\description{
Generate timeseries with changing bifurcation parameter
}
