% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cfun}
\alias{cfun}
\title{Utility function for foreach loop to save memory}
\usage{
cfun(...)
}
\arguments{
\item{...}{All outputs from foreach loop}
}
\description{
Utility function for foreach loop to save memory
}
