% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{combine_mixed_regimes}
\alias{combine_mixed_regimes}
\title{Combine regimes that are too short to constitute their own regime}
\usage{
combine_mixed_regimes(
  regimes_A,
  X_names,
  min_length_regime,
  name_mixed_regime = "Mixture: Periodic, Chaotic or Transitioning (X1,X2,X3,X4)",
  grepl_mixed_regime = function(regime) {
    
    grepl("Chaotic or Transitioning (X1,X2,X3,X4)", regime, fixed = TRUE) |
    (grepl("Chaotic or Transitioning", regime, fixed = TRUE) & grepl("Period", regime,
    fixed = TRUE))
 }
)
}
\arguments{
\item{regimes_A}{Dataframe with regime}

\item{X_names}{Names of variables}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}

\item{name_mixed_regime}{Name of new regime}

\item{grepl_mixed_regime}{Function to detect regimes that should be grouped together}
}
\value{
Dataframe with updated regimes
}
\description{
Combine regimes that are too short to constitute their own regime
}
