% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downsample}
\alias{downsample}
\title{Reduce size of dataframe by downsampling}
\usage{
downsample(
  df,
  X_names,
  win_size = 10,
  which_X = c("all", "first", "middle", "last", "random")[1],
  seed_nr = NULL
)
}
\arguments{
\item{df}{Dataframe}

\item{X_names}{Names of columns in df containing observational timeseries to be downsampled}

\item{win_size}{Number of data points to be averaged or sampled from in each step}

\item{which_X}{String of time point in window to select}

\item{seed_nr}{Seed number for random sampling of data point per window}
}
\value{
Downsampled dataframe
}
\description{
Reduce size of dataframe by downsampling
}
