% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{find_basin_boundary}
\alias{find_basin_boundary}
\title{Find whether and when basin boundaries are hit}
\usage{
find_basin_boundary(peaks_df, variable_name = "X1", min_edge = 0, max_edge = 1)
}
\arguments{
\item{peaks_df}{Dataframe with peaks}

\item{variable_name}{Column name in dataframe to assess}

\item{min_edge}{Minimum value reached}

\item{max_edge}{Maximum value reached}
}
\value{
Dataframe with bifurcation parameter value for which the system touches the basin boundaries as defined by min_edge and max_edge, if it hits these. Otherwise, dataframe filled with NA.
}
\description{
Find whether and when basin boundaries are hit
}
