% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{find_best_k}
\alias{find_best_k}
\title{Find best fitting period length k as determined by the minimum within cluster sum of squares}
\usage{
find_best_k(ks, coord, peak_idx)
}
\arguments{
\item{ks}{Vector of cluster sizes}

\item{coord}{Peak and trough coordinates}

\item{peak_idx}{Peak and trough indices}
}
\value{
Dataframe with best fitting period length k and the corresponding WCSS  for peak and trough coordinates and indices
}
\description{
Find best fitting period length k as determined by the minimum within cluster sum of squares
}
