% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{find_conseq_seq}
\alias{find_conseq_seq}
\title{Find consecutive values in sequence}
\usage{
find_conseq_seq(bifpar_idx_)
}
\arguments{
\item{bifpar_idx_}{Sequence of indices}
}
\value{
Start and end of each consecutive value sequence
}
\description{
Find consecutive values in sequence
}
