% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{find_expansion_basin_bound}
\alias{find_expansion_basin_bound}
\title{Update regimes with basin boundaries and chaotic expansion or reduction}
\usage{
find_expansion_basin_bound(
  peaks_df,
  regimes,
  min_length_regime,
  variable_name = "X1",
  min_edge = 0,
  max_edge = 1,
  thresh_expansion = 0.1
)
}
\arguments{
\item{peaks_df}{Dataframe with peaks}

\item{regimes}{Dataframe with regimes}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}

\item{variable_name}{Column name in dataframe to assess for hitting basin boundaries}

\item{min_edge}{Minimum basin boundary}

\item{max_edge}{Maximum basin boundary}

\item{thresh_expansion}{Threshold of difference in peak maxima at least one variable has to satisfy to qualify as expansion or reduction}
}
\value{
Updated dataframe
}
\description{
Update regimes with basin boundaries and chaotic expansion or reduction
}
