% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{find_regime_bounds}
\alias{find_regime_bounds}
\title{Find regime boundaries}
\usage{
find_regime_bounds(regimes, min_length_regime)
}
\arguments{
\item{regimes}{Dataframe with periodicity per value of the bifurcation parameter}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}
}
\value{
Dataframe with regime boundaries
}
\description{
Find regime boundaries
}
