% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_path}
\alias{format_path}
\title{Format and recursively create file path}
\usage{
format_path(pars_file)
}
\arguments{
\item{pars_file}{List of names of directories and filename components needed to create a directory}
}
\value{
File path
}
\description{
Format and recursively create file path
}
