% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_AUC}
\alias{get_AUC}
\title{Get Area Under the Curve (AUC)}
\usage{
get_AUC(fpr, tpr)
}
\arguments{
\item{fpr}{False Positive Rate}

\item{tpr}{True Positive Rate}
}
\value{
AUC
}
\description{
Get Area Under the Curve (AUC)
}
