% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_AUC_class}
\alias{get_AUC_class}
\title{Label AUC}
\usage{
get_AUC_class(AUC, nbins = 10)
}
\arguments{
\item{AUC}{Vector with Area Under the Curve}

\item{nbins}{Number of bins to divide AUC into}
}
\value{
Labelled AUC
}
\description{
Label AUC
}
