% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{get_bands}
\alias{get_bands}
\title{Get bifurcation diagram band properties}
\usage{
get_bands(x, min_edge = 0, max_edge = 1, step_size = 0.05)
}
\arguments{
\item{x}{Peak vector}

\item{min_edge}{Minimum basin boundary}

\item{max_edge}{Maximum basin boundary}

\item{step_size}{Step size in histogram}
}
\value{
Band properties
}
\description{
Get bifurcation diagram band properties
}
