% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_conn}
\alias{get_conn}
\title{Mean absolute cross-correlation}
\usage{
get_conn(x)
}
\arguments{
\item{x}{Matrix}
}
\value{
Mean absolute cross-correlation
}
\description{
Mean absolute cross-correlation
}
