% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_forloop}
\alias{get_forloop}
\title{Generate forloop for foreach loop}
\usage{
get_forloop(...)
}
\arguments{
\item{...}{Named forloop parameters}
}
\value{
List of forloop parameters
}
\description{
Generate forloop for foreach loop
}
