% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_formals}
\alias{get_formals}
\title{Get formal arguments of function as list}
\usage{
get_formals(func)
}
\arguments{
\item{func}{Function}
}
\value{
List with formal arguments and default arguments
}
\description{
Get formal arguments of function as list
}
