% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fs}
\alias{get_fs}
\title{Compute sampling frequency}
\usage{
get_fs(timestep = 0.01, sample_interval = 25)
}
\arguments{
\item{timestep}{Timestep}

\item{sample_interval}{Interval between samples}
}
\value{
Sampling frequency
}
\description{
Compute sampling frequency
}
