% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_Hurst_exp}
\alias{get_Hurst_exp}
\title{Hurst exponent}
\usage{
get_Hurst_exp(x, fs, nr_timesteps)
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}
}
\value{
Fitted linear slope of scale vs. detrended fluctuation (log-log)
}
\description{
Hurst exponent
}
