% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{get_regime_switch_type}
\alias{get_regime_switch_type}
\title{Find type of regime switch}
\usage{
get_regime_switch_type(from_regime, to_regime, X_names)
}
\arguments{
\item{from_regime}{The regime from which the transition occurs}

\item{to_regime}{The regime which is transitioned to}

\item{X_names}{Names of variables}
}
\value{
Regime switch type
}
\description{
Find type of regime switch
}
