% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_Smax}
\alias{get_Smax}
\title{Compute Maximum Spectral Density (Bury, 2021)}
\usage{
get_Smax(x, fs, nr_timesteps)
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}
}
\value{
Spectral EWS from Welch's PSD
}
\description{
Compute Maximum Spectral Density (Bury, 2021)
}
