% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_specEWS}
\alias{get_specEWS}
\title{Get EWS from spectral density}
\usage{
get_specEWS(x, fs, nr_timesteps, low = 0.05, high = 0.5, bw = 0.05)
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}

\item{low}{Low frequency}

\item{high}{High frequency}

\item{bw}{Width around low and high frequency for estimating spectral ratio}
}
\value{
Dataframe with spectral EWS
}
\description{
Get EWS from spectral density
}
