% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_spectral_exp}
\alias{get_spectral_exp}
\title{Spectral exponent}
\usage{
get_spectral_exp(x, fs, nr_timesteps)
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}
}
\value{
Slope of power spectral density
}
\description{
Spectral exponent
}
