% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_spectral_ratio}
\alias{get_spectral_ratio}
\title{Compute Spectral Ratio (Biggs, 2009)}
\usage{
get_spectral_ratio(
  x,
  fs,
  nr_timesteps,
  f_min_to_f_max = c(0.05, 0.5),
  n.freq = 5000
)
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}

\item{f_min_to_f_max}{Low and high frequency to compare}

\item{n.freq}{Number of points to estimate frequency at}
}
\value{
Spectral ratio of spectral density estimated at specified low frequency to spectral density at specified high frequency
}
\description{
Compute Spectral Ratio (Biggs, 2009)
}
