% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_spectral_ratio_exp}
\alias{get_spectral_ratio_exp}
\title{Spectral ratio of spectral density at low to high frequency}
\usage{
get_spectral_ratio_exp(x, fs, nr_timesteps, low = 0.05, high = 0.5)
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}

\item{low}{Low frequency}

\item{high}{High frequency}
}
\value{
Spectral ratio of spectral density at low frequency to spectral density at high frequency
}
\description{
Spectral ratio of spectral density at low to high frequency
}
