% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{get_stopifregime}
\alias{get_stopifregime}
\title{Define stopifregime function for bifurcation_ts()}
\usage{
get_stopifregime(
  type = c("default", "no_nodes", "only_nodes")[1],
  min_bifpar_idx = 2,
  max_bifpar_idx = 50
)
}
\arguments{
\item{type}{Type of stopifregime function; one of "default" (stop in case of infinite or NaN values), "no_nodes" (stop in case of a node), or "only_nodes" (stop in case of NO node)}

\item{min_bifpar_idx}{Minimum bifurcation parameter index above which stopifregime is applied}

\item{max_bifpar_idx}{Maximum bifurcation parameter index below which stopifregime is applied}
}
\value{
stopifregime function
}
\description{
Define stopifregime function for bifurcation_ts()
}
