% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_warnings_one_sigma}
\alias{get_warnings_one_sigma}
\title{Detect warnings based on a critical cut-off value}
\usage{
get_warnings_one_sigma(
  bifpar_idx,
  z_score,
  sigma_crit,
  nr_consecutive_warnings
)
}
\arguments{
\item{bifpar_idx}{Bifurcation parameter index}

\item{z_score}{Z-score}

\item{sigma_crit}{Critical cut-off value}

\item{nr_consecutive_warnings}{Number of consecutive warnings to look for}
}
\value{
Dataframe with warning indices and number of warning patches
}
\description{
Detect warnings based on a critical cut-off value
}
