% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_warnings_raw}
\alias{get_warnings_raw}
\title{Find warnings for cut-off value sigma-crit}
\usage{
get_warnings_raw(
  split_df_EWS,
  baseline_idx,
  transition_idx,
  sigma_crit = 2,
  nr_consecutive_warnings = 1
)
}
\arguments{
\item{split_df_EWS}{Dataframe of EWS split into lists, with one entry per value of the bifurcation parameter}

\item{baseline_idx}{Indices of baseline steps in bifurcation parameter}

\item{transition_idx}{Indices of transition steps in bifurcation parameter}

\item{sigma_crit}{Critical cut-off value}

\item{nr_consecutive_warnings}{Number of consecutive warnings to look for}
}
\value{
Dataframe with raw values of EWS as well as the confidence bands corresponding to sigma_crit; ready for plotting
}
\description{
Find warnings for cut-off value sigma-crit
}
