% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_Welch_specEWS}
\alias{get_Welch_specEWS}
\title{Spectral EWS from Welch's PSD}
\usage{
get_Welch_specEWS(
  x,
  fs,
  nr_timesteps,
  low = 0.005,
  high = 0.5,
  low_perc = 0.01,
  high_perc = 0.1,
  max_f = 50
)
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}

\item{low}{Low frequency}

\item{high}{High frequency}

\item{low_perc}{Percentage of low frequencies to sum for spectral ratio}

\item{high_perc}{Percentage of high frequencies to sum for spectral ratio}

\item{max_f}{Maximum frequency length to fit spectral slope to}
}
\value{
Spectral EWS from Welch's PSD
}
\description{
Spectral EWS from Welch's PSD
}
