% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_zscore_EWS}
\alias{get_zscore_EWS}
\title{Get z-scores of EWS}
\usage{
get_zscore_EWS(split_df_EWS, baseline_idx)
}
\arguments{
\item{split_df_EWS}{Dataframe of EWS split into lists, with one entry per value of the bifurcation parameter}

\item{baseline_idx}{Indices of baseline steps in bifurcation parameter}
}
\value{
Dataframe with the mean and standard deviation per EWS in the baseline period
}
\description{
Get z-scores of EWS
}
