% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQA.R
\name{getRM}
\alias{getRM}
\title{Thresholded Recurrence Matrix}
\usage{
getRM(
  x,
  emDim,
  emLag,
  theiler,
  distNorm,
  rescaleDist,
  targetValue,
  fix_emRad_or_RR = c("emRad-space-diameter", "emRad-constant", "RR")[2]
)
}
\arguments{
\item{x}{Data, can be a vector or matrix or dataframe}

\item{emDim}{Embedding dimension}

\item{emLag}{Embedding lag}

\item{theiler}{Size of Theiler window}

\item{distNorm}{Distance norm used for constructing recurrence matrix}

\item{rescaleDist}{Rescale distance matrix}

\item{targetValue}{Target value for thresholding recurrence matrix}

\item{fix_emRad_or_RR}{Set radius or recurrence rate as fixed}
}
\value{
Recurrence Matrix
}
\description{
Thresholded Recurrence Matrix
}
