% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLV_example.R
\docType{data}
\name{GLV_example}
\alias{GLV_example}
\title{Example data set of Generalized Lotka Volterra Model}
\format{
A data frame with {nrow} rows and {ncol} variables:
\describe{
\item{col1}{col1 description}
\item{col2}{col2 description}
}
}
\source{
\url{data_URL}
}
\usage{
GLV_example
}
\description{
Description of the dataset.
}
\keyword{datasets}
