% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{GLV_model}
\alias{GLV_model}
\title{Generalized Lotka-Volterra Model}
\usage{
GLV_model(t, state, parms)
}
\arguments{
\item{t}{Time}

\item{state}{Named vector of variable state}

\item{parms}{List of model parameters}
}
\value{
Updated state
}
\description{
Generalized Lotka-Volterra Model
}
