% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLV_SN_trans.R
\docType{data}
\name{GLV_SN_trans}
\alias{GLV_SN_trans}
\title{Saddle-node bifurcation in GLV model (transition model)}
\format{
A named list:
\describe{
\item{df}{Dataframe with timeseries}
\item{X_names}{Names of variables in model}
\item{X0s}{Dataframe with initial condition per bifurcation parameter value}
\item{timestep}{Step size in time}
\item{nr_timesteps}{Number of time steps}
\item{model_pars}{Fixed model parameters}
\item{seed_nr}{Seed number}
\item{stopifregime}{End generating timeseries if this function is satisfied}
\item{deSolve_method}{ODE in the style of deSolve}
\item{bifpar_list}{List with changing bifurcation parameter values}
\item{bifpar_pars}{List with parameters to generate bifpar_list, needs to include list(bifpar_start = .96, bifpar_end = .96), optionally specify list(pre_steps = 0, baseline_steps = 100, transition_steps = 0, post_steps = 0)}
\item{do_downsample}{Reduce dataframe size by downsampling?}
\item{fs}{Sampling frequency for downsampling if do_downsample = TRUE}
\item{which_X}{Time point(s) in downsampling window to select if do_downsample = TRUE}
}
}
\usage{
GLV_SN_trans
}
\description{
Data set of a saddle-node bifurcation in the Generalized Lotka Volterra model which was generated using bifurcation_ts()
}
\keyword{datasets}
