% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Lorenz_model}
\alias{Lorenz_model}
\title{Lorenz model}
\usage{
Lorenz_model(t, state, parms)
}
\arguments{
\item{t}{Time}

\item{state}{Named vector of variable state}

\item{parms}{List of model parameters}
}
\value{
Updated state
}
\description{
Lorenz model
}
