% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{make_filter_regime_switches}
\alias{make_filter_regime_switches}
\title{Make filter to obtain desired regime switches}
\usage{
make_filter_regime_switches(
  min_length_regime,
  baseline_steps,
  transition_steps
)
}
\arguments{
\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}

\item{baseline_steps}{Number of  steps for null model filter}

\item{transition_steps}{Number of transition steps for transition model filter}
}
\value{
List with filters for each desired regime switch. Each filter consists of:
regime_switch String; the name of the regime switch we're choosing here
}
\description{
Make filter to obtain desired regime switches
}
