% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{match_trans_null_model}
\alias{match_trans_null_model}
\title{Match transition and null model}
\usage{
match_trans_null_model(
  regime_bounds_trans,
  regime_bounds_null,
  min_length_regime,
  regime_switch_list,
  pre_steps,
  baseline_steps,
  transition_steps
)
}
\arguments{
\item{regime_bounds_trans}{Regime bounds of transition model, output of find_regimes()$regime_bounds}

\item{regime_bounds_null}{Regime bounds of null model, output of find_regimes()$regime_bounds}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}

\item{regime_switch_list}{List of filters per regime switch, output of make_filter_regime_switches()}

\item{pre_steps}{Number of pre-steps, useful if transition did not start when planned}

\item{baseline_steps}{Number of baseline steps where the control parameter doesn't change}

\item{transition_steps}{Number of transition steps where the control parameter changes for every index in a step-wise manner}
}
\value{
Dataframe with matched transition and null model.
}
\description{
Match transition and null model
}
