% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{max_dist}
\alias{max_dist}
\title{Find distance between points in chosen temporal sequence}
\usage{
max_dist(vec, cluster_idx)
}
\arguments{
\item{vec}{Vector}

\item{cluster_idx}{Vector of cluster indices}
}
\value{
Mean, minimum and maximum distance between vector points when partitioned using the indices in ks
}
\description{
Find distance between points in chosen temporal sequence
}
