% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_list}
\alias{modify_list}
\title{Modify list}
\usage{
modify_list(old_list, new_list)
}
\arguments{
\item{old_list}{List to modify}

\item{new_list}{List to modify old_list with}
}
\value{
Modified list
}
\description{
Modify list
}
