% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_downsample_pars}
\alias{new_downsample_pars}
\title{Update downsampling parameters with new sampling frequency}
\usage{
new_downsample_pars(downsample_pars, timestep = 0.01, new_fs = 5)
}
\arguments{
\item{downsample_pars}{List of parameters for downsampling}

\item{timestep}{Timestep}

\item{new_fs}{New sampling frequency}
}
\value{
List of updated parameters for downsampling
}
\description{
Update downsampling parameters with new sampling frequency
}
