% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{peaks_bifdiag}
\alias{peaks_bifdiag}
\title{Find both local peaks and troughs of timeseries}
\usage{
peaks_bifdiag(df, X_names)
}
\arguments{
\item{df}{Dataframe}

\item{X_names}{Column names of df to find peaks and troughs for}
}
\value{
Dataframe with peaks and troughs per variable, including index of the beginning, height, and end of each peak
}
\description{
Find both local peaks and troughs of timeseries
}
