% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{periods_to_regimes}
\alias{periods_to_regimes}
\title{Update periodicities with basin boundaries and merged bands}
\usage{
periods_to_regimes(
  peaks_df,
  periods,
  X_names,
  min_length_regime,
  variable_name = "X1",
  min_edge = 0,
  max_edge = 1,
  thresh_full_band = 0.8,
  nr_smooth = 0,
  factor_nr_steps = 3
)
}
\arguments{
\item{peaks_df}{Dataframe with peaks}

\item{periods}{Dataframe with periodicities}

\item{X_names}{Names of variables}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}

\item{variable_name}{Column name in dataframe to assess for hitting basin boundaries}

\item{min_edge}{Minimum edge of basin boundary}

\item{max_edge}{Maximum edge of basin boundary}

\item{thresh_full_band}{Percentage of bins that need to be occupied to qualify it as a fully filled banned}

\item{nr_smooth}{Number of exceptions in a stable periodicity window to smooth over; nr_smooth = 0 means no smoothing}

\item{factor_nr_steps}{Dividing factor for number of steps in histogram to determine merged bands}
}
\value{
Updated dataframe
}
\description{
Update periodicities with basin boundaries and merged bands
}
