% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_bifdiag}
\alias{plot_bifdiag}
\title{Plot bifurcation diagram}
\usage{
plot_bifdiag(
  regime_list,
  sel_variables = NULL,
  filepath_image = NULL,
  col_palette_regimes = c("lapaz", "batlowW")[2],
  alpha_regime = 0.99,
  sz_point_peaks = 1e-04,
  resolution = 200,
  factor_resolution = 3,
  reverse_x_axis = F,
  nbreaks_x = 8
)
}
\arguments{
\item{regime_list}{List with detected regimes, output of find_regimes()}

\item{sel_variables}{Selected variables to plot}

\item{filepath_image}{Filepath to save image to; if not specified, no plot is saved.}

\item{col_palette_regimes}{Name of colour palette to colour regimes. Must be one of scico's palettes, see all options with scico::scico_palette_names()}

\item{alpha_regime}{Transparancy of regime colours}

\item{sz_point_peaks}{Point size of peaks}

\item{resolution}{Image resolution}

\item{factor_resolution}{Factor to multiply image size by to increase resolution}

\item{reverse_x_axis}{Logical indicating whether the x-axis should be reversed}

\item{nbreaks_x}{Number of ticks on the x-axis}
}
\value{
Styled bifurcation digram
}
\description{
Plot bifurcation diagram
}
