% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{rev_scale_range}
\alias{rev_scale_range}
\title{Reverse scaling of vector between range (a,b)}
\usage{
rev_scale_range(x, a, b, min_x, max_x)
}
\arguments{
\item{x}{Vector}

\item{a}{Minimum value}

\item{b}{Maximum value}

\item{min_x}{Previous minimum of x}

\item{max_x}{Previous maximum of x}
}
\value{
Scaled vector
}
\description{
Reverse scaling of vector between range (a,b)
}
