% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQA.R
\name{rp_calc_KCE}
\alias{rp_calc_KCE}
\title{rp_calc}
\usage{
rp_calc_KCE(
  RM,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  WLmin = 1,
  DLmax = length(Matrix::diag(RM)),
  VLmax = length(Matrix::diag(RM)),
  HLmax = length(Matrix::diag(RM)),
  WLmax = length(Matrix::diag(RM)),
  theiler = NA,
  AUTO = NULL,
  includeDiagonal = NA,
  matrices = FALSE
)
}
\arguments{
\item{RM}{A binarized distance matrix}

\item{DLmin}{Minimal diagonal line length (default = \code{2})}

\item{VLmin}{Minimal vertical line length (default = \code{2})}

\item{HLmin}{Minimal horizontal line length (default = \code{2})}

\item{WLmin}{Minimal white line length (default = \code{1}) # KCE}

\item{DLmax}{Maximal diagonal line length (default = length of diagonal -1)}

\item{VLmax}{Maximal vertical line length (default = length of diagonal -1)}

\item{HLmax}{Maximal horizontal line length (default = length of diagonal -1)}

\item{WLmax}{Maximal white line length (default = length of diagonal -1) # KCE}

\item{theiler}{Theiler window}

\item{AUTO}{Auto-recurrence?}

\item{includeDiagonal}{Should the diagonal be included in the calculation of the Recurrence Rate? If \code{NA} the value will be decided by the symmetry of the matrix, the diagonal will be removed for Auto RQA (\code{AUTO = TRUE}) but not for Cross RQA (\code{AUTO = FALSE})  (default = \code{NA})}

\item{matrices}{Return matrices? (default = \code{FALSE})}
}
\value{
CRQA measures and matrices of line distributions (if requested)
}
\description{
rp_calc
}
