% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQA.R
\name{rp_measures_KCE}
\alias{rp_measures_KCE}
\title{Get RQA measures based on a binary matrix}
\usage{
rp_measures_KCE(
  RM,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  WLmin = 1,
  DLmax = length(Matrix::diag(RM)),
  VLmax = length(Matrix::diag(RM)),
  HLmax = length(Matrix::diag(RM)),
  WLmax = length(Matrix::diag(RM)),
  AUTO = NULL,
  theiler = NA,
  includeDiagonal = NA,
  matrices = FALSE,
  targetValue = 0.05,
  silent = TRUE
)
}
\arguments{
\item{RM}{A binarized distance matrix}

\item{DLmin}{Minimal diagonal line length (default = \code{2})}

\item{VLmin}{Minimal vertical line length (default = \code{2})}

\item{HLmin}{Minimal horizontal line length (default = \code{2})}

\item{WLmin}{Minimal white line length (default = \code{1}) # KCE}

\item{DLmax}{Maximal diagonal line length (default = length of diagonal -1)}

\item{VLmax}{Maximal vertical line length (default = length of diagonal -1)}

\item{HLmax}{Maximal horizontal line length (default = length of diagonal -1)}

\item{WLmax}{Maximal white line length (default = length of diagonal -1) # KCE}

\item{AUTO}{Auto-recurrence?}

\item{theiler}{Theiler window}

\item{includeDiagonal}{Should the diagonal be included in the calculation of the Recurrence Rate? If \code{NA} the value will be decided by the symmetry of the matrix, the diagonal will be removed for Auto RQA (\code{AUTO = TRUE}) but not for Cross RQA (\code{AUTO = FALSE})  (default = \code{NA})}

\item{matrices}{Return matrices? (default = \code{FALSE})}

\item{targetValue}{A value passed to \code{est_radius(...,type="fixed", targetMeasure="RR", tol = .2)} if \code{is.na(emRad)==TRUE}, it will estimate a radius (default = \code{.05}).}

\item{silent}{Do not display any messages (default = \code{TRUE})'}
}
\description{
A zoo of measures based on singular recurrent points, diagonal, vertical, horizontal, and white line structures (anisotropic) will be caluclated.
}
