% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{run_bifEWS}
\alias{run_bifEWS}
\title{Run EWS for all bifurcation parameter values}
\usage{
run_bifEWS(
  df,
  X_names,
  uni_metrics = c(Smax = get_Smax),
  multi_metrics = c(RQA = runRQA),
  EWS_args = list(Smax = list(fs = 1, nr_timesteps = 100), RQA = list(emDim = 1, emLag =
    1, theiler = 1, distNorm = "max", targetValue = 0.05))
)
}
\arguments{
\item{df}{Complete dataframe with timeseries for all bifurcation parameter values}

\item{X_names}{Names of variables in model}

\item{uni_metrics}{List of univariate EWS}

\item{multi_metrics}{List of multivariate EWS}

\item{EWS_args}{List of parameters to pass to EWS functions, empty list if no parameters need to be passed}
}
\value{
Dataframe with EWS
}
\description{
Run EWS for all bifurcation parameter values
}
