% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{runRQA}
\alias{runRQA}
\title{Recurrence Quantification Analysis}
\usage{
runRQA(
  x,
  emDim = 1,
  emLag = 1,
  theiler = 1,
  distNorm = "max",
  targetValue = 0.05
)
}
\arguments{
\item{x}{Data, can be a vector or matrix or dataframe}

\item{emDim}{Embedding dimension}

\item{emLag}{Embedding lag}

\item{theiler}{Size of Theiler window}

\item{distNorm}{Distance norm used for constructing recurrence matrix}

\item{targetValue}{Target value for thresholding recurrence matrix}
}
\value{
RQA measures
}
\description{
Recurrence Quantification Analysis
}
