% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{Utility function to save plots}
\usage{
save_plot(
  pl,
  filepath_image,
  w = 10,
  h = 10,
  resolution = 200,
  formats = c("tiff", "png", "pdf")[1]
)
}
\arguments{
\item{pl}{ggplot object}

\item{filepath_image}{Filepath image needs to be saved to}

\item{w}{Width}

\item{h}{Height}

\item{resolution}{Resolution}

\item{formats}{File formats}
}
\value{
Success of saving file
}
\description{
Utility function to save plots
}
