% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{scale_range}
\alias{scale_range}
\title{Scale vector between range (a,b)}
\usage{
scale_range(x, a = 0, b = 1)
}
\arguments{
\item{x}{Vector}

\item{a}{Minimum value}

\item{b}{Maximum value}
}
\value{
Scaled vector
}
\description{
Scale vector between range (a,b)
}
