% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_bifpars}
\alias{setup_bifpars}
\title{Specify parameters for specific regime switches}
\usage{
setup_bifpars()
}
\value{
List with parameters per regime switch
}
\description{
Specify parameters for specific regime switches
}
