% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_pars}
\alias{setup_pars}
\title{Set up parameters}
\usage{
setup_pars(model_name, pars_add = list())
}
\arguments{
\item{model_name}{Chosen dynamical systems model}
}
\value{
List of parameters
}
\description{
Set up parameters
}
