% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{smooth_column}
\alias{smooth_column}
\title{Smooth over exceptions in otherwise stable regime}
\usage{
smooth_column(
  rough_df,
  col_to_smooth = "period",
  smooth_along_col = c(),
  nr_smooth = 0,
  min_length_regime = 10
)
}
\arguments{
\item{rough_df}{Dataframe with periodicity per bifurcation parameter}

\item{col_to_smooth}{Column to smooth over}

\item{smooth_along_col}{Columns to smooth along too}

\item{nr_smooth}{Number of exceptions in a stable periodicity window to smooth over; nr_smooth = 0 means no smoothing}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}
}
\value{
Dataframe with smoothed periodicity
}
\description{
Smooth over exceptions in otherwise stable regime
}
