% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{smooth_periods}
\alias{smooth_periods}
\title{Smooth over exceptions in otherwise stable regime}
\usage{
smooth_periods(periods, nr_smooth, min_length_regime)
}
\arguments{
\item{periods}{Dataframe with periodicity per bifurcation parameter}

\item{nr_smooth}{Number of exceptions in a stable periodicity window to smooth over}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}
}
\value{
Dataframe with smoothed periodicity
}
\description{
Smooth over exceptions in otherwise stable regime
}
