% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{solve_sampling_par}
\alias{solve_sampling_par}
\title{Solve for sampling parameters}
\usage{
solve_sampling_par(fs = NULL, sample_interval = NULL, timestep = NULL)
}
\arguments{
\item{fs}{Sampling frequency}

\item{sample_interval}{Sampling interval}

\item{timestep}{Timestep}
}
\value{
List of sampling parameters
}
\description{
Solve for sampling parameters
}
