% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{style_plot}
\alias{style_plot}
\title{Style plot}
\usage{
style_plot(
  pl,
  col_pal = c(col_facet_labels = (scales::viridis_pal(option = "rocket", direction =
    -1))(20)[17]),
  fs = c(family = "serif", strip.text.x = 12, strip.text.y = 12, plot.title = 20,
    plot.subtitle = 16, axis.text = 8, axis.title = 16, legend.text = 12, legend.title =
    14, legend.spacing.y = 0.075)
)
}
\arguments{
\item{pl}{Original plot}

\item{col_pal}{Named colours}

\item{fs}{Font sizes}
}
\value{
Styled plot
}
\description{
Style plot
}
