% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{WCSS}
\alias{WCSS}
\title{Compute within cluster sum of squares (WCSS)}
\usage{
WCSS(vec, cluster_idx)
}
\arguments{
\item{vec}{Vector to partition into temporal clusters}

\item{cluster_idx}{Vector with cluster assignment of each point in vec, such as c(1,2,3,1,2,3) for k = 3 clusters}
}
\value{
Within cluster sum of squares
}
\description{
Compute within cluster sum of squares (WCSS)
}
