% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{YoudensJ}
\alias{YoudensJ}
\title{Compute best threshold using Youden's J statistic}
\usage{
YoudensJ(x)
}
\arguments{
\item{x}{Dataframe with true/false positive/negative rates for each critical cut-off value}
}
\value{
Dataframe with only the row corresponding to the maximal Youden's J statistic
}
\description{
Compute best threshold using Youden's J statistic
}
