% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{bifurcation_ts}
\alias{bifurcation_ts}
\title{Generate timeseries with changing bifurcation parameter}
\usage{
bifurcation_ts(
  model,
  model_pars,
  bifpar_list,
  X0 = c(),
  X_names = names(X0),
  seed_nr = 123,
  times = seq(0, 100, by = 0.01),
  deSolve_method = c("lsoda", "euler", "rk4")[1],
  stopifregime = function(out) {
     FALSE
 },
  do_downsample = TRUE,
  downsample_pars = list(type = c("average", "one_sample")[1], win_size = 10, which_X =
    c(10, "first", "middle", "last", "random")[1], seed = 123),
  silent = FALSE
)
}
\arguments{
\item{model}{ODE in the style of deSolve}

\item{model_pars}{Fixed model parameters}

\item{bifpar_list}{List with changing bifurcation parameter values}

\item{X0}{Initial state}

\item{X_names}{Names of variables in model}

\item{seed_nr}{Seed number}

\item{times}{Vector of timepoints}

\item{deSolve_method}{Method of generating ODE passed to deSolve::ode}

\item{stopifregime}{End generating timeseries if this function is satisfied}
}
\value{
List with timeseries matrix and parameters used to generate the timeseries
}
\description{
Generate timeseries with changing bifurcation parameter
}
