% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{choose_best_k}
\alias{choose_best_k}
\title{Choose best period length}
\usage{
choose_best_k(
  spread_df,
  thresh_node,
  factor_k,
  method_best_k = c("log", "min_scaled")[2]
)
}
\arguments{
\item{spread_df}{Output of find_spread_per_k()}

\item{thresh_node}{Threshold under which timeseries is classified as node}

\item{factor_k}{Weighting of period length k; heavier weight means shorter k is preferred; factor_k = 0 means the optimal period length is chosen based solely on minimum spread}

\item{method_best_k}{Method of choosing best period length}
}
\value{
Dataframe containing one row with best period length
}
\description{
Choose best period length
}
