% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downsample}
\alias{downsample}
\title{Reduce size of dataframe by downsampling}
\usage{
downsample(
  df,
  X_names,
  type = c("average", "one_sample")[1],
  win_size = 10,
  which_X = c(10, "first", "middle", "last", "random")[1],
  seed_nr = 123
)
}
\arguments{
\item{df}{Dataframe}

\item{X_names}{Names of columns in df containing observational timeseries to be downsampled}

\item{type}{Type of downsampling. "average" averages data points in each window; "one_sample" picks one sample in each window.}

\item{win_size}{Number of data points to be averaged or sampled from in each step}

\item{which_X}{Time point(s) in window to select}

\item{seed_nr}{Seed number for random sampling of data point per window}
}
\value{
Downsampled dataframe.
}
\description{
Reduce size of dataframe by downsampling
}
