% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{find_best_k}
\alias{find_best_k}
\title{Find best fitting period length k}
\usage{
find_best_k(coord, peak_idx, thresh_node, max_k = NULL)
}
\arguments{
\item{coord}{Peak and trough coordinates}

\item{peak_idx}{Peak and trough indices}

\item{thresh_node}{Threshold under which timeseries is classified as node}

\item{max_k}{Maximum cluster size to look for}
}
\value{
Dataframe with best fitting period length k and the corresponding minimum within-cluster distance and between-cluster distance for peak and trough coordinates and indices
}
\description{
Find best fitting period length k
}
