% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{find_regimes}
\alias{find_regimes}
\title{Find regimes of dynamical system}
\usage{
find_regimes(
  df,
  X_names,
  thresh_node = 0.1,
  thresh_coord_spread = 0.025,
  thresh_peak_idx_spread = 2,
  min_length_regime = 5,
  nr_smooth = 2,
  max_k = NULL
)
}
\arguments{
\item{df}{Dataframe}

\item{X_names}{Names of variables in model}

\item{thresh_node}{Threshold under which timeseries is classified as node}

\item{thresh_coord_spread}{Threshold for distance in peak and trough coordinates that determines whether a cluster (i.e. period length) fits; if exceeded, denoted as chaotic or transitioning}

\item{thresh_peak_idx_spread}{Same as thresh_coord_spread but for peak and trough indices}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}

\item{nr_smooth}{Number of exceptions in a stable periodicity window to smooth over}

\item{max_k}{Maximum cluster size to look for}
}
\value{
List of dataframes with periodicity per variable, periodicity per bifurcation parameter value, regimes, and regime boundaries
}
\description{
Find regimes of dynamical system
}
