% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{find_regimes}
\alias{find_regimes}
\title{Find regimes of dynamical system}
\usage{
find_regimes(
  df,
  X_names,
  ks = 2:100,
  thresh_coord_WCSS = 0.2,
  thresh_peak_idx_WCSS = 0.2,
  min_length_regime = 10
)
}
\arguments{
\item{df}{Dataframe}

\item{X_names}{Names of variables in model}

\item{ks}{Sequence of cluster sizes, starts at 2}

\item{thresh_coord_WCSS}{Threshold for within cluster sum of squares in peak and trough coordinates that determines a cluster (i.e. period length) fits; if exceeded, denoted as chaotic or transitioning}

\item{thresh_peak_idx_WCSS}{Same as thresh_coord_WCSS but for peak and trough indices}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}
}
\value{
List of dataframes with periodicity per variable, periodicity per bifurcation parameter value, regimes, and regime boundaries
}
\description{
Find regimes of dynamical system
}
