% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{get_bands}
\alias{get_bands}
\title{Get bifurcation diagram band properties}
\usage{
get_bands(x, min_x = 0, max_x = 1, nr_steps = 10)
}
\arguments{
\item{x}{Peak vector}

\item{min_x}{Minimum x}

\item{max_x}{Maximum x}

\item{nr_steps}{Number of steps in histogram}
}
\value{
Band properties
}
\description{
Get bifurcation diagram band properties
}
