% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_Smax}
\alias{get_Smax}
\title{Get maximum spectral density}
\usage{
get_Smax(x, fs, nr_timesteps)
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}
}
\value{
Maximum spectral density and corresponding frequency
}
\description{
Get maximum spectral density
}
