% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_spectral_exp}
\alias{get_spectral_exp}
\title{Compute Spectral Exponent (Prettyman, 2020; Wijnants, 2013)}
\usage{
get_spectral_exp(
  x,
  fs,
  nr_timesteps,
  method = c("Prettyman2020", "Wijnants2013")[2]
)
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}

\item{method}{Method of selecting frequency range in which to compute spectral exponent}
}
\value{
Slope of power spectral density
}
\description{
Compute Spectral Exponent (Prettyman, 2020; Wijnants, 2013)
}
