% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_spectral_exp}
\alias{get_spectral_exp}
\title{Compute Spectral Exponent (Prettyman, 2020)}
\usage{
get_spectral_exp(x, fs, nr_timesteps, f_min = 10^(-2), f_max = 10^(-1))
}
\arguments{
\item{x}{Signal}

\item{fs}{Sampling frequency}

\item{nr_timesteps}{Number of timesteps}

\item{f_min}{Low frequency}

\item{f_max}{High frequency}
}
\value{
Slope of power spectral density
}
\description{
Compute Spectral Exponent (Prettyman, 2020)
}
