% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_warnings}
\alias{get_warnings}
\title{Get warnings per EWS metric}
\usage{
get_warnings(
  split_df_EWS,
  baseline_steps,
  transition_steps,
  sigmas_crit = seq(0.25, 6, by = 0.25),
  nr_consecutive_warnings = 1
)
}
\arguments{
\item{split_df_EWS}{Dataframe of EWS split into lists, with one entry per value of the bifurcation parameter}

\item{baseline_steps}{Number of baseline steps in bifurcation parameter}

\item{transition_steps}{Number of transition steps in bifurcation parameter}

\item{sigmas_crit}{Sequence of critical values of sigma}

\item{nr_consecutive_warnings}{Number of consecutive warnings to look for}
}
\value{
Dataframe with warnings per EWS metric
}
\description{
Get warnings per EWS metric
}
