% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{get_warnings_per_sigma}
\alias{get_warnings_per_sigma}
\title{Get warnings per critical value of sigma}
\usage{
get_warnings_per_sigma(
  y,
  bifpar_idx,
  z_score,
  sigma_crit_step,
  thresh_max_sigma_crit = 500,
  nr_consecutive_warnings = 1
)
}
\arguments{
\item{y}{Vector with EWS}

\item{bifpar_idx}{Bifurcation parameter index}

\item{z_score}{Z-score}

\item{sigma_crit_step}{Step size in sequence of critical values of sigma}

\item{thresh_max_sigma_crit}{Maximum critical cut-off value to look for}

\item{nr_consecutive_warnings}{Number of consecutive warnings to look for}
}
\value{
Dataframe with warnings
}
\description{
Get warnings per critical value of sigma
}
