% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{periods_to_regimes}
\alias{periods_to_regimes}
\title{Update periodicities with basin boundaries}
\usage{
periods_to_regimes(
  peaks_df,
  periods,
  X_names,
  min_length_regime,
  variable_name = "X1",
  min_edge = 0,
  max_edge = 1
)
}
\arguments{
\item{peaks_df}{Dataframe with peaks}

\item{periods}{Dataframe with periodicities}

\item{X_names}{Names of variables}

\item{min_length_regime}{Minimum number of consecutive steps in the bifurcation parameter that have the same periodicity to qualify as a regime}

\item{variable_name}{Column name in dataframe to assess for hitting basin boundaries}

\item{min_edge}{Minimum basin boundary}

\item{max_edge}{Maximum basin boundary}
}
\value{
Updated dataframe
}
\description{
Update periodicities with basin boundaries
}
