% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_3D_landscape}
\alias{plot_3D_landscape}
\title{Plot 3D landscape}
\usage{
plot_3D_landscape(
  df,
  bifpar_name = "s",
  bifpar_idx = 1,
  bifpar_value = 0.6,
  plot_idx = 1,
  colors = c("red"),
  size_marker = 1,
  size_line = 0.1,
  plot_mode = c("markers", "lines+markers")[2]
)
}
\arguments{
\item{df}{Dataframe}

\item{bifpar_name}{Name of bifurcation parameter}

\item{bifpar_idx}{Index of bifurcation parameter}

\item{bifpar_value}{Value of bifurcation parameter}

\item{plot_idx}{Index of subplot}

\item{colors}{Vector of colors}

\item{size_marker}{Size of marker}

\item{size_line}{Linewidth}

\item{plot_mode}{Plotting mode}
}
\value{
Plotly object
}
\description{
Plot 3D landscape
}
