% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_3D_landscape_trans}
\alias{plot_3D_landscape_trans}
\title{Plot transition in 3D}
\usage{
plot_3D_landscape_trans(
  df,
  bifpar_idxs,
  bifpar_list,
  nrows = NULL,
  ncols = NULL,
  x1_range = c(0, 1),
  x2_range = c(0, 1),
  x3_range = c(0, 1),
  size_marker = 1,
  size_line = 0.1,
  plot_mode = c("markers", "lines+markers")[2]
)
}
\arguments{
\item{df}{Dataframe}

\item{bifpar_idxs}{Bifurcation parameter indices to plot}

\item{bifpar_list}{List of bifurcation parameters to index}

\item{nrows}{Number of rows}

\item{ncols}{Number of columns}

\item{x1_range}{Range of x1-axis}

\item{x2_range}{Range of x2-axis}

\item{x3_range}{Range of x3-axis}

\item{size_marker}{Size of marker}

\item{size_line}{Linewidth}

\item{plot_mode}{Plotting mode}
}
\value{
Plotly object
}
\description{
Plot transition in 3D
}
