% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{run_EWS}
\alias{run_EWS}
\title{Compute early warning signals (EWS)}
\usage{
run_EWS(df, uni_metrics, multi_metrics, pars = list())
}
\arguments{
\item{df}{Dataframe}

\item{uni_metrics}{List of univariate EWS}

\item{multi_metrics}{List of multivariate EWS}

\item{pars}{List of parameters to pass to EWS functions, empty list if no parameters need to be passed}
}
\value{
Dataframe with early warning signals
}
\description{
Compute early warning signals (EWS)
}
