% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWS.R
\name{runRQA}
\alias{runRQA}
\title{Recurrence Quantification Analysis}
\usage{
runRQA(
  x,
  emDim = 1,
  emLag = 1,
  theiler = 1,
  distNorm = "max",
  targetValue = 0.05
)
}
\arguments{
\item{targetValue}{}
}
\value{
RQA measures
}
\description{
Recurrence Quantification Analysis
}
