% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{smooth_periods}
\alias{smooth_periods}
\title{Smooth over exceptions in otherwise stable regime}
\usage{
smooth_periods(periods, min_length_regime)
}
\arguments{
\item{periods}{Dataframe with periodicity per bifurcation parameter}

\item{min_length_regime}{Minimum length of regime}
}
\value{
Dataframe with smoothed periodicity
}
\description{
Smooth over exceptions in otherwise stable regime
}
