% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{style_plot}
\alias{style_plot}
\title{Style plot}
\usage{
style_plot(pl, col_pal = c(col_facet_labels = "#3C1A42FF"), fs = c())
}
\arguments{
\item{pl}{Original plot}

\item{col_pal}{Named colours}

\item{fs}{Font sizes}
}
\value{
Styled plot
}
\description{
Style plot
}
